+++ NEWS.md
@@ -1,11 +1,16 @@
-# boxed 0.1.1 (dev version)
+# boxes 0.2.0 (dev version)
+
+* `pack`: use object name as default `id`.
+* Rename package to `boxes`.
+
+# boxed 0.1.1
 
 * `remove`, `item_remove`: remove an item from a box.
-* rename package to `boxed`. A depot becomes a `box`, `keep` becomes `pack`
+* Rename package to `boxed`. A depot becomes a `box`, `keep` becomes `pack`
 
 # keeper 0.1.0
 
 * `depot_export` and `depot_import`: store and retrieve a depot in a file on/from disk.
 * `keep_file` to add a file from disk to a depot.
 * `keep` and `pick` to add and retrieve objects from depot.
-* add basic depot handlers `depots_*` and `depot_*`.
+* Add basic depot handlers `depots_*` and `depot_*`.